Initial draft of void ParticleFilter::prediction
should consider that the prediction algorithm may really be based on theta_dot, velocity and delta_t which have a std_distribution which is different than the GPS coordinates.  We are using previous GPS coords to update the prediction step which when initialized had GPS sensor error included so the next iteration shouldn't be adding the same error.  Consider creating
  // vehicle sensor input uncertainty [u [m/s], dt [s], theta_dot [rad/s]]
  double sigma_sense [3] = {0.03, 0.000000001, 0.0001};